Render markdown response and move video preview into previews folder
