+++ src/renderer/jobs/JobViewOutputs.tsx
@@ -1,7 +1,9 @@
 import { useParams } from 'react-router-dom';
 import PreviewModal from 'src/renderer/components/PreviewModal';
 import videoResponse from 'src/shared/dummy_data/file_response_video';
+import markdownResponseBody from 'src/shared/dummy_data/markdown_response';
 import { useJob } from '../lib/hooks';
+import MarkdownView from '../components/response_body/MarkdownView';
 
 function JobViewOutputs() {
   const { jobId } = useParams();
@@ -16,8 +18,8 @@ function JobViewOutputs() {
   if (!response) return <div>No response available</div>;
 
   return (
-    <div>
-      <div>Job ID: {jobId}</div>
+    <div className="border border-gray-300 rounded-lg mt-5 p-4 shadow-md bg-white">
+      <MarkdownView response={markdownResponseBody} />
       <PreviewModal response={videoResponse} />
     </div>
   );
